requirements.txt
Django==4.2.1
djangorestframework==3.14.0
black==23.3.0
psycopg2-binary==2.9.6
